Made GUI abstract as well
